requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
isort==5.5.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
yapf==0.30.0
